+++ __mocks__/chat-message.ts
@@ -1,0 +1,4 @@
+// @ts-ignore
+global.ChatMessage = {
+    create: jest.fn().mockResolvedValue(null)
+};